api/requirements.txt
amqp==5.2.0
async-timeout==4.0.3
billiard==4.2.0
boto3==1.28.84
botocore==1.31.84
celery==5.3.5
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==2.2.2
flower==2.0.1
greenlet==3.0.1
humanize==4.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.3
MarkupSafe==2.1.3
numpy==1.24.3
prometheus-client==0.18.0
prompt-toolkit==3.0.40
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.23
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
Werkzeug==2.2.2
